Lazy load images
That works better than I expected. What remains is getting optimal compression
on the images, and loading different sizes based on the viewport width the
further conserve mobile data.

It may also be a good idea to host the lazyload js locally and concat our
index.js with it to save an extra request.